+++ gitlab-ci/packer/gitlab_image.json
@@ -1,0 +1,15 @@
+{
+    "builders": [
+        {
+            "type": "yandex",
+            "service_account_key_file": "{{ user `key_file` }}",
+            "folder_id": "{{ user `folder_id` }}",
+            "source_image_family": "{{ user `image` }}",
+            "image_name": "gitlab-ci",
+            "image_family": "gitlab-ci",
+            "ssh_username": "ubuntu",
+            "platform_id": "standard-v1",
+            "use_ipv4_nat": "true"
+        }
+    ]
+}
