+++ gitlab-ci/reddit/views/show.haml
@@ -1,0 +1,45 @@
+%div{ id: 'postlist'}
+  %div{class: 'panel'}
+    %div{class: 'panel-heading'}
+      %div{class: 'text-center'}
+        %div{class: 'row'}
+          .col-sm-1
+            %form{action: "#{@post['_id']['$oid']}/vote/1", method: "post"}
+              %input{:type => "hidden", :name => "_method", :value => "put"}
+              %button{type: "submit", class: "btn btn-default btn-sm"}
+                %span{ class: "glyphicon glyphicon-menu-up" }
+            %h4{class: 'pull-center'} #{@post['votes']}
+            %form{:action => "#{@post['_id']['$oid']}/vote/-1", :method => "post"}
+              %input{:type => "hidden", :name => "_method", :value=> "put"}
+              %button{type: "submit", class: "btn btn-default btn-sm"}
+                %span{ class: "glyphicon glyphicon-menu-down" }
+          .col-sm-8
+            %h3{class: 'pull-left'}
+              %a{href: "#{@post['_id']['$oid']}"} #{@post['title']}
+          .col-sm-3
+            %h4{class: 'pull-right'}
+              %small
+                %em #{Time.at(@post['created_at'].to_i).strftime('%d-%m-%Y')}
+                %br #{Time.at(@post['created_at'].to_i).strftime('%H:%M')}
+    .panel-footer
+      %a{href: "#{@post['link']}", class: 'btn btn-link'} Go to the link
+
+- unless @comments.nil? or @comments.empty?
+  - @comments.each do |comment|
+    .row
+      .col-sm-8
+        .panel.panel-default
+          .panel-heading
+            %strong #{comment['name']}
+            %span{class: "text-muted pull-right"}
+              %em #{Time.at(comment['created_at'].to_i).strftime('%H:%M')} #{Time.at(comment['created_at'].to_i).strftime('%d-%m-%Y')}
+          .panel-body #{comment['body']}
+- if logged_in?
+  %form{action: "/post/#{@post['_id']['$oid']}/comment", method: 'post', role: 'form'}
+    .col-sm-8
+      .form-group
+        %textarea{name: 'body', class: 'form-control', placeholder: 'put a nice comment :)'}
+      %button{class: 'btn btn-block btn-primary'} Post my comment
+- else
+  .alert{class: "alert-info"}
+    %strong Log in or sign up to add a new comment! :)
