+++ gitlab-ci/terraform/.terraform.lock.hcl
@@ -1,0 +1,9 @@
+# This file is maintained automatically by "terraform init".
+# Manual edits may be lost in future updates.
+
+provider "registry.terraform.io/yandex-cloud/yandex" {
+  version = "0.77.0"
+  hashes = [
+    "h1:SSv4dS9ifKx8TaXIpQ/2PRgwLImcJBDXv+O2Fz+Gg8Q=",
+  ]
+}
