+++ gitlab-ci/terraform/main.tf
@@ -1,0 +1,21 @@
+terraform {
+  required_providers {
+    yandex = {
+      source = "yandex-cloud/yandex"
+    }
+  }
+  required_version = ">= 0.13"
+}
+
+#Для подключения к облаку
+provider "yandex" {
+  service_account_key_file = var.service_account_key_file
+  cloud_id                 = var.cloud_id
+  folder_id                = var.folder_id
+  zone                     = var.zone
+}
+
+/* module "vpc" { */
+/*   source = "./modules/vpc" */
+/* } */
+
