+++ gitlab-ci/terraform/modules/vpc/outputs.tf
@@ -1,0 +1,3 @@
+output "app_subnet_id" {
+  value = yandex_vpc_subnet.app-subnet.id
+}
