+++ gitlab-ci/terraform/outputs.tf
@@ -1,0 +1,6 @@
+output "external_ip_address_app" {
+  value = yandex_compute_instance.app[*].network_interface.0.ip_address
+}
+output "internal_ip_address_app" {
+  value = yandex_compute_instance.app[*].network_interface.0.nat_ip_address
+}
