+++ gitlab-ci/terraform/terraform.tfstate
@@ -1,0 +1,187 @@
+{
+  "version": 4,
+  "terraform_version": "1.2.6",
+  "serial": 15,
+  "lineage": "6a2f97ef-1c20-2d12-63e0-2077285ce44c",
+  "outputs": {
+    "external_ip_address_app": {
+      "value": [
+        "192.168.10.3"
+      ],
+      "type": [
+        "tuple",
+        [
+          "string"
+        ]
+      ]
+    },
+    "internal_ip_address_app": {
+      "value": [
+        "51.250.69.199"
+      ],
+      "type": [
+        "tuple",
+        [
+          "string"
+        ]
+      ]
+    }
+  },
+  "resources": [
+    {
+      "mode": "managed",
+      "type": "yandex_compute_instance",
+      "name": "app",
+      "provider": "provider[\"registry.terraform.io/yandex-cloud/yandex\"]",
+      "instances": [
+        {
+          "schema_version": 1,
+          "attributes": {
+            "allow_recreate": null,
+            "allow_stopping_for_update": null,
+            "boot_disk": [
+              {
+                "auto_delete": true,
+                "device_name": "fhm86ks1joipghgjj7gp",
+                "disk_id": "fhm86ks1joipghgjj7gp",
+                "initialize_params": [
+                  {
+                    "block_size": 4096,
+                    "description": "",
+                    "image_id": "fd8vpsdq3f8pqrfrf8ls",
+                    "name": "",
+                    "size": 10,
+                    "snapshot_id": "",
+                    "type": "network-hdd"
+                  }
+                ],
+                "mode": "READ_WRITE"
+              }
+            ],
+            "created_at": "2022-08-21T14:19:58Z",
+            "description": "",
+            "folder_id": "b1gmcu0esncmdden02bk",
+            "fqdn": "fhmh1fm82h0bq7sic91i.auto.internal",
+            "hostname": "",
+            "id": "fhmh1fm82h0bq7sic91i",
+            "labels": {
+              "tags": "gitlab-ci"
+            },
+            "local_disk": [],
+            "metadata": {
+              "ssh-keys": "ubuntu:ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQC96NTxAi7XycfSSzHJSvd3AYxA63/hFWwB+UgzC1mk43Km8DhPZgLFLWN6Z9dn1w4umEqfp2xv5Z2Mm7ItY5/71w2It2gr1aDM7vSKs5RoW7qiNvtcJIXN5eNMTu6uRYuACa/znh6PMg7hFA3qAWGXh+tn8qd88+G19fd4Oac4oB7sb6Hz6AwSZeRQiyJY7y4LhuxRDN9vMGfnh3BB9p3fBgCNDi9j1w+WtuM6As0YP1d945ezmyjaPw9rCffIJ0/P/pqgbYkU8oDFczFVE5Na6xH3rgivYQ3fazXk04MctVvm0JB070x3ePWqOZN4CDRYmw1AxwTTN/haQCwL0TUhxwazhZ8x1OgAcHWmwE2052uyADxJuJQkByvUyR6aK5tWs2bMUyUi+9hA3mBInT+I6Azfi+S5fwiaQ/co8yhqwna+erAAmUV8E3CElGacwtZz1Sv8+U1q4pg1lJUPuKRvuNwx2FADL4oHo398YkEuqBjUJRjYlWLj0CT84LJN27k= appuser\n"
+            },
+            "name": "gitlab-ci",
+            "network_acceleration_type": "standard",
+            "network_interface": [
+              {
+                "dns_record": [],
+                "index": 0,
+                "ip_address": "192.168.10.3",
+                "ipv4": true,
+                "ipv6": false,
+                "ipv6_address": "",
+                "ipv6_dns_record": [],
+                "mac_address": "d0:0d:11:0b:ec:81",
+                "nat": true,
+                "nat_dns_record": [],
+                "nat_ip_address": "51.250.69.199",
+                "nat_ip_version": "IPV4",
+                "security_group_ids": [],
+                "subnet_id": "e9b020eu7cj3qpl83666"
+              }
+            ],
+            "placement_policy": [
+              {
+                "host_affinity_rules": [],
+                "placement_group_id": ""
+              }
+            ],
+            "platform_id": "standard-v1",
+            "resources": [
+              {
+                "core_fraction": 100,
+                "cores": 2,
+                "gpus": 0,
+                "memory": 4
+              }
+            ],
+            "scheduling_policy": [
+              {
+                "preemptible": true
+              }
+            ],
+            "secondary_disk": [],
+            "service_account_id": "",
+            "status": "running",
+            "timeouts": null,
+            "zone": "ru-central1-a"
+          },
+          "sensitive_attributes": [],
+          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiY3JlYXRlIjozMDAwMDAwMDAwMDAsImRlbGV0ZSI6MzAwMDAwMDAwMDAwLCJ1cGRhdGUiOjMwMDAwMDAwMDAwMH0sInNjaGVtYV92ZXJzaW9uIjoiMSJ9",
+          "dependencies": [
+            "yandex_vpc_network.app-network",
+            "yandex_vpc_subnet.app-subnet"
+          ]
+        }
+      ]
+    },
+    {
+      "mode": "managed",
+      "type": "yandex_vpc_network",
+      "name": "app-network",
+      "provider": "provider[\"registry.terraform.io/yandex-cloud/yandex\"]",
+      "instances": [
+        {
+          "schema_version": 0,
+          "attributes": {
+            "created_at": "2022-08-21T14:19:56Z",
+            "default_security_group_id": "",
+            "description": "",
+            "folder_id": "b1gmcu0esncmdden02bk",
+            "id": "enp4idfa9muisg8d8352",
+            "labels": {},
+            "name": "app-network",
+            "subnet_ids": [],
+            "timeouts": null
+          },
+          "sensitive_attributes": [],
+          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiY3JlYXRlIjo2MDAwMDAwMDAwMCwiZGVsZXRlIjo2MDAwMDAwMDAwMCwidXBkYXRlIjo2MDAwMDAwMDAwMH19"
+        }
+      ]
+    },
+    {
+      "mode": "managed",
+      "type": "yandex_vpc_subnet",
+      "name": "app-subnet",
+      "provider": "provider[\"registry.terraform.io/yandex-cloud/yandex\"]",
+      "instances": [
+        {
+          "schema_version": 0,
+          "attributes": {
+            "created_at": "2022-08-21T14:19:57Z",
+            "description": "",
+            "dhcp_options": [],
+            "folder_id": "b1gmcu0esncmdden02bk",
+            "id": "e9b020eu7cj3qpl83666",
+            "labels": {},
+            "name": "app-subnet",
+            "network_id": "enp4idfa9muisg8d8352",
+            "route_table_id": "",
+            "timeouts": null,
+            "v4_cidr_blocks": [
+              "192.168.10.0/24"
+            ],
+            "v6_cidr_blocks": [],
+            "zone": "ru-central1-a"
+          },
+          "sensitive_attributes": [],
+          "private": "eyJlMmJmYjczMC1lY2FhLTExZTYtOGY4OC0zNDM2M2JjN2M0YzAiOnsiY3JlYXRlIjoxODAwMDAwMDAwMDAsImRlbGV0ZSI6MTgwMDAwMDAwMDAwLCJ1cGRhdGUiOjE4MDAwMDAwMDAwMH19",
+          "dependencies": [
+            "yandex_vpc_network.app-network"
+          ]
+        }
+      ]
+    }
+  ]
+}
