+++ gitlab-ci/terraform/terraform.tfstate.backup
@@ -1,0 +1,8 @@
+{
+  "version": 4,
+  "terraform_version": "1.2.6",
+  "serial": 11,
+  "lineage": "6a2f97ef-1c20-2d12-63e0-2077285ce44c",
+  "outputs": {},
+  "resources": []
+}
