+++ gitlab-ci/terraform/terraform.tfvars
@@ -1,0 +1,11 @@
+service_account_key_file = "/home/zemlyak/.ssh/key_terraform.json"
+token                    = "AQAAAABiPeLEAATuwSaAY2x3ykJ_jmEb6nKswt4"
+cloud_id                 = "b1ght3nmtm711gtojq4e"
+folder_id                = "b1gmcu0esncmdden02bk"
+zone                     = "ru-central1-a"
+# subnet_id                = "e9bsgtde700a5reltcc6"
+gitlab_image          = "fd8vpsdq3f8pqrfrf8ls"
+public_key_path         = "~/.ssh/appuser.pub"
+private_key_path        = "~/.ssh/appuser"
+/* access_key              = "YCAJET_8w1VJx1DDQ8k4k2GDy" */
+/* secret_key              = "YCPRrY2AZ54fMWiVDDDP5NURp55WQSJw9DA85mM5" */
