--- src/gitlab/reddit/Gemfile
@@ -1,17 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'sinatra', '~> 2.0.1'
-gem 'haml'
-gem 'bson_ext'
-gem 'bcrypt'
-gem 'puma'
-gem 'mongo'
-gem 'json'
-gem 'rack-test'
-
-group :development do
-    gem 'capistrano',         require: false
-    gem 'capistrano-rvm',     require: false
-    gem 'capistrano-bundler', require: false
-    gem 'capistrano3-puma',   require: false
-end
