--- src/gitlab/reddit/config.ru
@@ -1,2 +1,0 @@
-require './app'
-run Sinatra::Application
