--- src/gitlab/reddit/config/deploy.rb
@@ -1,69 +1,0 @@
-server ENV['SERVER_IP'], port: 22, user: ENV['DEPLOY_USER'], roles: [:web, :app, :db], primary: true
-set :repo_name, ENV['REPO_NAME']
-
-set :application, 'reddit'
-set :repo_url, "git@github.com:#{fetch(:repo_name)}.git"
-set :branch, 'monolith'
-set :user, ENV['DEPLOY_USER']
-set :puma_threads,    [4, 16]
-set :puma_workers,    0
-
-set :pty,             true
-set :use_sudo,        false
-set :stage,           :production
-set :deploy_via,      :remote_cache
-set :deploy_to,       "/home/#{fetch(:user)}/#{fetch(:application)}"
-set :puma_bind,       "tcp://0.0.0.0:9292"
-set :puma_state,      "#{shared_path}/tmp/pids/puma.state"
-set :puma_pid,        "#{shared_path}/tmp/pids/puma.pid"
-set :puma_access_log, "#{release_path}/log/puma.error.log"
-set :puma_error_log,  "#{release_path}/log/puma.access.log"
-set :ssh_options,     { forward_agent: true }
-set :puma_preload_app, true
-set :puma_worker_timeout, nil
-set :puma_init_active_record, false  # Change to false when not using ActiveRecord
-set :linked_dirs, %w(tmp/pids tmp/sockets log)
-
-namespace :puma do
-  desc 'Create Directories for Puma Pids and Socket'
-  task :make_dirs do
-    on roles(:app) do
-      execute "mkdir #{shared_path}/tmp/sockets -p"
-      execute "mkdir #{shared_path}/tmp/pids -p"
-    end
-  end
-
-  before :start, :make_dirs
-end
-
-namespace :deploy do
-  desc "Make sure local git is in sync with remote."
-  task :check_revision do
-    on roles(:app) do
-      unless `git rev-parse HEAD` == `git rev-parse origin/monolith`
-        puts "WARNING: HEAD is not the same as origin/monolith"
-        puts "Run `git push` to sync changes."
-        exit
-      end
-    end
-  end
-
-  desc 'Initial Deploy'
-  task :initial do
-    on roles(:app) do
-      before 'deploy:restart', 'puma:start'
-      invoke 'deploy'
-    end
-  end
-
-  desc 'Restart application'
-  task :restart do
-    on roles(:app), in: :sequence, wait: 5 do
-      invoke 'puma:restart'
-    end
-  end
-
-  before :starting,     :check_revision
-  after  :finishing,    :cleanup
-  after  :finishing,    :restart
-end
