--- src/gitlab/reddit/helpers.rb
@@ -1,29 +1,0 @@
-helpers do
-  # a helper method to turn a string ID
-  # representation into a BSON::ObjectId
-  def object_id val
-    begin
-      BSON::ObjectId.from_string(val)
-    rescue BSON::ObjectId::Invalid
-      nil
-    end
-  end
-
-  def document_by_id id
-    id = object_id(id) if String === id
-    if id.nil?
-      {}.to_json
-    else
-      document = settings.mongo_db.find(:_id => id).to_a.first
-      (document || {}).to_json
-    end
-  end
-
-  def logged_in?
-      if session[:username].nil?
-          return false
-      else
-          return true
-      end
-  end
-end
