--- src/gitlab/reddit/simpletest.rb
@@ -1,18 +1,0 @@
-require_relative './app'
-require 'test/unit'
-require 'rack/test'
-
-set :environment, :test
-
-class MyAppTest < Test::Unit::TestCase
-  include Rack::Test::Methods
-
-  def app
-    Sinatra::Application
-  end
-
-  def test_get_request
-    get '/'
-    assert last_response.ok?
-  end
-end
