--- src/gitlab/reddit/views/create.haml
@@ -1,14 +1,0 @@
-%h2 Add a new post
-- if logged_in?
-  %form{ action: '/new', method: 'post', role: 'form'}
-    .form-group
-      %label{for: 'title'} Title:
-      %input{name:'title', placeholder: 'cool title', class: 'form-control', id: 'title'}
-    .form-group
-      %label{for: 'link'} Link:
-      %input{name:'link', placeholder: 'awesome link', class: 'form-control', id: 'link'}
-    .form-group
-      %input{class: 'btn btn-primary', type: 'submit', value:'Post it!'}
-- else
-  .alert{class: "alert-danger"}
-    %strong You need to sign in before adding a new post :(
