--- src/gitlab/reddit/views/layout.haml
@@ -1,46 +1,0 @@
-!!! 5
-%html(lang="en")
-  %head
-    %meta(charset="utf-8")
-    %meta(http-equiv="X-UA-Compatible" content="IE=Edge,chrome=1")
-    %meta(name="viewport" content="width=device-width, initial-scale=1.0")
-    %title="Monolith Reddit :: #{@title}"
-    %link{ href: 'https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css', integrity: "sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7", type: 'text/css', rel: 'stylesheet', crossorigin: 'anonymous' }
-    %link{ href: 'https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css', integrity: 'sha384-fLW2N01lMqjakBkx3l/M9EahuwpSfeNvV63J5ezn3uZzapT0u7EYsXMjQV+0En5r', type: 'text/css', rel: 'stylesheet', crossorigin: 'anonymous' }
-    %script{ href: 'https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js', integrity: 'sha384-0mSbJDEHialfmuBBQP6A4Qrprq5OVfW37PRR3j5ELqxss1yVqOtnepnHVP9aJ7xS', crossorigin: 'anonymous'}
-    %script{ src: 'https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js' }
-  %body
-    .navbar.navbar-default.navbar-static-top
-      .container
-        .navbar-header
-          %button.navbar-toggle(type="button" data-toggle="collapse" data-target=".navbar-responsive-collapse")
-            %span.icon-bar
-            %span.icon-bar
-            %span.icon-bar
-          %a.navbar-brand(href="/") Monolith Reddit
-        .navbar-collapse.collapse
-          %ul.nav.navbar-nav.navbar-right
-            - if logged_in?
-              %li
-                %a.nav-link(href="/logout") Logout
-            - else
-              %li
-                %a(href="/signup") Sign up
-              %li
-                %a(href="/login") Login
-    .container
-      .row
-        .col-lg-9
-          - unless @flashes.nil? or @flashes.empty?
-            - @flashes.each do |flash|
-              .alert{class: flash[:type]}
-                %strong #{flash[:message]}
-          = yield
-        .col-lg-3
-          .well.sidebar-nav
-            %h3 Menu
-            %ul.nav.nav-list
-              %li
-                %a(href="/") All posts
-              %li
-                %a(href="/new") New post
